backend/requirements.txt
django
djangorestframework
django-filter
django-cors-headers
psycopg2-binary
gunicorn
requests
python-dotenv
